Returns Rule ARN when describing Rule from Remote account
This commit prevents an issue when trying to describe rules created
on EventBuses from Remote Accounts.

This change returns the ARN of the Rule, that the Remote Account would
have to use reference it, as 'unqualified' names refer to Rules attached
to the local account EventBus.

We only do this when describing from Remote accounts, rather than
returning the full ARN back because it would remain compatible with the
majority of deployed resources. Returning the ARN to rules created on
the local bus would cause many plans and changes.